Move netlify cms specific content extraction into netlify_cms folder
