+++ apps/codeforafrica/content/projects/civicsignal.md
@@ -1,5 +1,6 @@
 ---
 id: 7e3a1ef1-576d-4428-824d-a12fa774519d
+slug: civic-signal
 name: CivicSignal
 tagLine: Africa media ecosystem
 title: |
@@ -19,17 +20,15 @@ thumbnail:
   src: https://res.cloudinary.com/code-for-africa/image/upload/v1652705959/codeforafrica/images/Property_1_PesaCheck_iahlrh.jpg
 href: /projects/civic-signal
 externalHref: https://codeforafrica.org
-partners:
-  - daa2a622-0277-44be-b302-dbc703f89522
-  - e90b7786-69b3-4694-9ee9-d9c21dc80287
-  - 1079f9ae-0fbd-472b-bbfb-518145c28b1d
-slug: civic-signal
 badges:
   id:
     - 5e6d2d87-6819-4aee-a08c-e2c6b8ef7226
     - cce4b68a-ad5e-428e-a9aa-849a8a7afb27
+partners:
+  - daa2a622-0277-44be-b302-dbc703f89522
+  - e90b7786-69b3-4694-9ee9-d9c21dc80287
+  - 1079f9ae-0fbd-472b-bbfb-518145c28b1d
 donors:
-  - f67426f2-9338-43cc-85c3-0782420eb623
   - f690fb82-9baa-4e31-a4b2-d57ffad43874
 team:
   - 8f56e20e-eb6c-44f2-be0a-cf96e75a91fa
