+++ apps/codeforafrica/src/lib/netlify_cms/getCollectionData.js
