+++ apps/codeforafrica/src/lib/netlify_cms/getPartners.js
