+++ apps/codeforafrica/src/lib/netlify_cms/getProjects.js
